requirements.txt
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv
llama-index-llms-openai
llama-index-embeddings-openai
